sdk/samples/EPRI dataset import/requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
diffgram==0.8.5
idna==3.3
imageio==2.19.5
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
scipy==1.8.1
six==1.16.0
urllib3==1.26.10
